Extract triggerLayerFeatureEvent function
