+++ src/api/v4/native/leaflet-layer.js
@@ -2,6 +2,7 @@ var L = require('leaflet');
 var _ = require('underscore');
 var Layer = require('../layer');
 var constants = require('../constants');
+var triggerLayerFeatureEvent = require('./trigger-layer-feature-event');
 var LeafletCartoLayerGroupView = require('../../../geo/leaflet/leaflet-cartodb-layer-group-view');
 
 /**
@@ -65,7 +66,7 @@ var LeafletLayer = L.TileLayer.extend({
   },
 
   _onFeatureClick: function (internalEvent) {
-    this._triggerLayerFeatureEvent(Layer.events.FEATURE_CLICKED, internalEvent);
+    triggerLayerFeatureEvent(Layer.events.FEATURE_CLICKED, internalEvent, this._layers);
   },
 
   _onFeatureOver: function (internalEvent) {
@@ -74,7 +75,7 @@ var LeafletLayer = L.TileLayer.extend({
       this._hoveredLayers[internalEvent.layerIndex] = true;
       this._map.getContainer().style.cursor = 'pointer';
     }
-    this._triggerLayerFeatureEvent(Layer.events.FEATURE_OVER, internalEvent);
+    triggerLayerFeatureEvent(Layer.events.FEATURE_OVER, internalEvent, this._layers);
   },
 
   _onFeatureOut: function (internalEvent) {
@@ -84,44 +85,7 @@ var LeafletLayer = L.TileLayer.extend({
     } else {
       this._map.getContainer().style.cursor = 'auto';
     }
-    this._triggerLayerFeatureEvent(Layer.events.FEATURE_OUT, internalEvent);
-  },
-
-  _triggerLayerFeatureEvent: function (eventName, internalEvent) {
-    var layer = this._layers.findById(internalEvent.layer.id);
-    if (layer) {
-      var event = {
-        data: undefined,
-        latLng: undefined
-      };
-      if (internalEvent.feature) {
-        event.data = internalEvent.feature;
-      }
-      if (internalEvent.latlng) {
-        event.latLng = {
-          lat: internalEvent.latlng[0],
-          lng: internalEvent.latlng[1]
-        };
-      }
-      if (internalEvent.position) {
-        event.position = {
-          x: internalEvent.position.x,
-          y: internalEvent.position.y
-        };
-      }
-
-      /**
-       *
-       * Events triggered by {@link carto.layer.Layer} when users interact with a feature.
-       *
-       * @event carto.layer.Layer.FeatureEvent
-       * @property {LatLng} latLng - Object with coordinates where interaction took place
-       * @property {object} data - Object with feature data (one attribute for each specified column)
-       *
-       * @api
-       */
-      layer.trigger(eventName, event);
-    }
+    triggerLayerFeatureEvent(Layer.events.FEATURE_OUT, internalEvent, this._layers);
   }
 });
